+++ source/renderer/app/actions/wallets-actions.js
@@ -1,7 +1,7 @@
 // @flow
 import Action from './lib/Action';
 import type { WalletExportTypeChoices } from '../types/walletExportTypes';
-import type { CsvRecord } from '../../../common/types/rewards-csv-request.types';
+import type { CsvFileContent } from '../../../common/types/csv-request.types';
 
 export type WalletImportFromFileParams = {
   filePath: string,
@@ -57,15 +57,21 @@ export default class WalletsActions {
     walletExportType: WalletExportTypeChoices,
   }> = new Action();
   generateCertificate: Action<{ filePath: string }> = new Action();
-  generateRewardsCsv: Action<{
+  generateCsv: Action<{
     filePath: string,
-    rewards: Array<CsvRecord>,
+    fileContent: CsvFileContent,
   }> = new Action();
   generateAddressPDF: Action<{
-    address: string,
     note: string,
+    address: string,
     filePath: string,
   }> = new Action();
+  generateAddressPDFSuccess: Action<{ walletAddress: string }> = new Action();
+  saveQRCodeImage: Action<{
+    address: string,
+    filePath: string,
+  }> = new Action();
+  saveQRCodeImageSuccess: Action<{ walletAddress: string }> = new Action();
   copyAddress: Action<{ address: string }> = new Action();
   updateCertificateStep: Action<any> = new Action();
   closeCertificateGeneration: Action<any> = new Action();