+++ source/renderer/app/components/staking/delegation-setup-wizard/DelegationStepsChooseWalletDialog.js
@@ -80,6 +80,12 @@ const messages = defineMessages({
     description:
       'Label for continue button on the delegation setup "choose wallet" step dialog.',
   },
+  syncingWallet: {
+    id: 'staking.delegationSetup.chooseWallet.step.dialog.syncingWallet',
+    defaultMessage: '!!!syncing',
+    description:
+      'Syncing wallet label on the delegation setup "choose wallet" step dialog.',
+  },
 });
 
 type Props = {
@@ -228,6 +234,7 @@ export default class DelegationStepsChooseWalletDialog extends Component<
               messages.selectWalletInputPlaceholder
             )}
             value={selectedWalletId}
+            syncingLabel={intl.formatMessage(messages.syncingWallet)}
             getStakePoolById={getStakePoolById}
           />
           {error}