+++ source/renderer/app/components/staking/stake-pools/StakePoolsRanking.js
@@ -1,12 +1,17 @@
 // @flow
 import React, { Component } from 'react';
 import { observer } from 'mobx-react';
-import { defineMessages, intlShape } from 'react-intl';
+import { defineMessages, intlShape, FormattedHTMLMessage } from 'react-intl';
 import classnames from 'classnames';
+import { PopOver } from 'react-polymorph/lib/components/PopOver';
 import { ButtonSkin } from 'react-polymorph/lib/skins/simple/ButtonSkin';
-import { TooltipSkin } from 'react-polymorph/lib/skins/simple/TooltipSkin';
-import { Tooltip } from 'react-polymorph/lib/components/Tooltip';
-import { shortNumber, generateThousands } from '../../../utils/formatters';
+import BigNumber from 'bignumber.js';
+import {
+  shortNumber,
+  generateThousands,
+  formattedWalletAmount,
+  toFixedUserFormat,
+} from '../../../utils/formatters';
 import {
   getFilteredWallets,
   getAllAmounts,
@@ -49,7 +54,7 @@ const messages = defineMessages({
   rankingDescription: {
     id: 'staking.stakePools.rankingDescription',
     defaultMessage:
-      '!!!Use the slider to rank the stake pools based on the amount you intend to delegate.',
+      '!!!Use the slider to rank the stake pools and check the potential rewards <strong>based on the amount of stake you intend to delegate</strong>.',
     description: 'Ranking description.',
   },
   rankingLearnMoreUrl: {
@@ -126,7 +131,7 @@ type Props = {
 
 type State = {
   sliderValue: number,
-  amountValue: number,
+  displayValue: string,
 };
 
 @observer
@@ -143,15 +148,19 @@ export default class StakePoolsRanking extends Component<Props, State> {
     sliderValue: Math.round(
       INITIAL_DELEGATION_FUNDS_LOG * RANKING_SLIDER_RATIO
     ),
-    amountValue: INITIAL_DELEGATION_FUNDS,
+    displayValue: toFixedUserFormat(INITIAL_DELEGATION_FUNDS, 0),
   };
 
   componentDidMount() {
     const { stake } = this.props;
     if (stake) {
+      const hasDecimal = stake - Math.floor(stake);
+      const displayValue = hasDecimal
+        ? formattedWalletAmount(new BigNumber(stake), false)
+        : toFixedUserFormat(stake, 0);
       this.setState({
         sliderValue: Math.round(Math.log(stake) * RANKING_SLIDER_RATIO),
-        amountValue: stake,
+        displayValue,
       });
     }
   }
@@ -174,17 +183,17 @@ export default class StakePoolsRanking extends Component<Props, State> {
     const selectedWallet = wallets.find(
       (wallet) => wallet.id === selectedWalletId
     );
+    const hasSelectedWallet = !!selectedWallet;
+    const isAllWalletsSelected = selectedWalletId === ALL_WALLETS_SELECTION_ID;
+    const wasSelectedWalletChanged =
+      selectedWalletId !== selectedDelegationWalletId;
 
-    if (
-      selectedWalletId === selectedDelegationWalletId ||
-      (selectedWalletId !== ALL_WALLETS_SELECTION_ID && !selectedWallet)
-    ) {
+    // Prevent ranking stake pools if we don't have data for the selected wallet ready
+    if (wasSelectedWalletChanged && !isAllWalletsSelected && !hasSelectedWallet)
       return;
-    }
 
     let amountValue = 0;
     let sliderValue = 0;
-
     if (selectedWalletId === ALL_WALLETS_SELECTION_ID) {
       amountValue = Math.min(
         getAllAmounts(wallets).toNumber(),
@@ -195,7 +204,16 @@ export default class StakePoolsRanking extends Component<Props, State> {
     }
     amountValue = Math.max(amountValue, MIN_DELEGATION_FUNDS);
     sliderValue = Math.round(Math.log(amountValue) * RANKING_SLIDER_RATIO);
-    this.setState({ sliderValue, amountValue });
+    const hasSliderValueChanged = sliderValue !== this.state.sliderValue;
+
+    // Prevent ranking stake pools if selected wallet and slider value remains unchanged
+    if (!wasSelectedWalletChanged && !hasSliderValueChanged) return;
+
+    const displayValue = formattedWalletAmount(
+      new BigNumber(amountValue),
+      false
+    );
+    this.setState({ sliderValue, displayValue });
     updateDelegatingStake(selectedWalletId, amountValue);
     rankStakePools();
   };
@@ -218,7 +236,8 @@ export default class StakePoolsRanking extends Component<Props, State> {
         Math.exp(sliderValue / RANKING_SLIDER_RATIO)
       );
     }
-    this.setState({ sliderValue, amountValue });
+    const displayValue = toFixedUserFormat(amountValue, 0);
+    this.setState({ sliderValue, displayValue });
     updateDelegatingStake(null, amountValue);
   };
 
@@ -278,8 +297,7 @@ export default class StakePoolsRanking extends Component<Props, State> {
       getStakePoolById,
       rankStakePools,
     } = this.props;
-    const { sliderValue, amountValue } = this.state;
-    const rankingDescription = intl.formatMessage(messages.rankingDescription);
+    const { sliderValue, displayValue } = this.state;
     const learnMoreButtonClasses = classnames(['flat', styles.actionLearnMore]);
     const {
       walletSelectorWallets,
@@ -298,7 +316,9 @@ export default class StakePoolsRanking extends Component<Props, State> {
         <div className={styles.upper}>
           <div className={styles.selectWallet}>
             <div className={styles.row}>
-              <div className={styles.col}>{rankingDescription}</div>
+              <div className={styles.col}>
+                <FormattedHTMLMessage {...messages.rankingDescription} />
+              </div>
             </div>
             {getFilteredWallets(wallets).length > 0 ? (
               <div className={styles.row}>
@@ -366,7 +386,7 @@ export default class StakePoolsRanking extends Component<Props, State> {
                 )}
                 maxDisplayValue={MAX_DELEGATION_FUNDS}
                 value={sliderValue}
-                displayValue={amountValue}
+                displayValue={displayValue}
                 showRawValue
                 onChange={this.onSliderChange}
                 onAfterChange={rankStakePools}
@@ -378,12 +398,11 @@ export default class StakePoolsRanking extends Component<Props, State> {
             </div>
             <div className={styles.col}>
               <div className={styles.outOfRangeMaxAmount}>
-                <Tooltip
-                  skin={TooltipSkin}
-                  tip={intl.formatMessage(messages.rankingExtraTooltip)}
+                <PopOver
+                  content={intl.formatMessage(messages.rankingExtraTooltip)}
                 >
                   {shortNumber(OUT_OF_RANGE_MAX_DELEGATION_FUNDS)}
-                </Tooltip>
+                </PopOver>
               </div>
               <div className={styles.outOfSliderRangeEnd} />
             </div>