+++ source/renderer/app/components/staking/widgets/ThumbPoolContent.js
@@ -15,15 +15,21 @@ import {
 
 type Props = {
   stakePool: StakePool,
-  numberOfStakePools: number,
+  numberOfRankedStakePools: number,
 };
 
 @observer
 export default class ThumbPoolContent extends Component<Props> {
   render() {
-    const { stakePool, numberOfStakePools } = this.props;
-    const { ranking, ticker, retiring, saturation } = stakePool;
-    const color = getColorFromRange(ranking, numberOfStakePools);
+    const { stakePool, numberOfRankedStakePools } = this.props;
+    const {
+      ranking,
+      nonMyopicMemberRewards,
+      ticker,
+      retiring,
+      saturation,
+    } = stakePool;
+    const color = getColorFromRange(ranking, numberOfRankedStakePools);
 
     const componentClassnames = classnames([
       styles.component,
@@ -39,19 +45,32 @@ export default class ThumbPoolContent extends Component<Props> {
       <div className={componentClassnames}>
         <div className={styles.ticker}>{ticker}</div>
         {IS_RANKING_DATA_AVAILABLE ? (
-          <>
-            <div className={styles.ranking} style={{ color }}>
-              {ranking}
-            </div>
-            {IS_SATURATION_DATA_AVAILABLE && (
-              <div className={saturationClassnames}>
-                <span
-                  style={{
-                    width: `${parseFloat(saturation.toFixed(2))}%`,
-                  }}
-                />
-              </div>
+          <div className={styles.ranking} style={{ color }}>
+            {nonMyopicMemberRewards ? (
+              ranking
+            ) : (
+              <>
+                {numberOfRankedStakePools + 1}
+                <sup>*</sup>
+              </>
             )}
+          </div>
+        ) : (
+          <div className={styles.noDataDash}>
+            <SVGInline svg={noDataDashBigImage} />
+          </div>
+        )}
+        {IS_SATURATION_DATA_AVAILABLE && (
+          <div className={saturationClassnames}>
+            <span
+              style={{
+                width: `${parseFloat(saturation).toFixed(2)}%`,
+              }}
+            />
+          </div>
+        )}
+        {IS_RANKING_DATA_AVAILABLE ? (
+          <>
             {retiring && (
               <div className={styles.clock}>
                 <SVGInline svg={clockIcon} className={styles.clockIcon} />
@@ -65,12 +84,7 @@ export default class ThumbPoolContent extends Component<Props> {
             />
           </>
         ) : (
-          <>
-            <div className={styles.noDataDash}>
-              <SVGInline svg={noDataDashBigImage} />
-            </div>
-            <div className={styles.greyColorBand} />
-          </>
+          <div className={styles.greyColorBand} />
         )}
       </div>
     );